backend/requirements.txt
# Framework Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
hiredis==2.3.2

# AI & ML
openai==1.10.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
langchain==0.1.6
langchain-openai==0.0.5

# PDF Processing
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.2

# Email
resend==0.7.0

# Validation & Serialization
pydantic==2.6.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Utils
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14